Renaming static_panel:kind to :encoding
#1316
