+++ waltz-ng/client/static-panel/components/static-panel.html
