+++ waltz-ng/client/static-panel/components/static-panels.html
